spring-example/backend/src/main/java/com/hamza/customer/CustomerJPADataAccessService.java
package com.hamza.customer;

import org.springframework.stereotype.Repository;
import java.util.List;
import java.util.Optional;

@Repository("jpa")
class CustomerJPADataAccessService implements CustomerDao{
    private final CustomerRepository customerRepository;

    CustomerJPADataAccessService(CustomerRepository customerRepository) {
        this.customerRepository = customerRepository;
    }

    @Override
    public List<Customer> selectAllCustomers() {
        return customerRepository.findAll();
    }

    @Override
    public Optional<Customer> selectCustomerById(long id) {
        return customerRepository.findById(id);
    }

    @Override
    public void addCustomer(Customer customer) {
        customerRepository.save(customer);
    }

    @Override
    public boolean customerExistsByEmail(String email) {
        return customerRepository.existsByEmail(email);
    }

    @Override
    public void removeCustomerById(long id) {
        customerRepository.deleteById(id);
    }

    @Override
    public void updateCustomer(Customer customer) {
        customerRepository.save(customer);
    }
}
